Crud prof - gestion message d'erreur

+++ controllers/matieres.controller.js
@@ -1,4 +1,5 @@
 const { ObjectId } = require("bson");
+const assignment = require("../model/assignment");
 
 let Matiere = require("../model/matiere");
 
@@ -22,13 +23,17 @@ function getMatieres(req, res) {
             .then(matieres => {
                 res.send(matieres);
             }).catch(err => {
-                res.send(err);
+                res.status(500).json({
+                    message: "Une erreur est survenue, veuillez réessayer"
+                })
             });
     } else {
         aggregateQuery.then(matieres => {
             res.send(matieres);
         }).catch(err => {
-            res.send(err);
+            res.status(500).json({
+                message: "Une erreur est survenue, veuillez réessayer"
+            })
         });
     }
 }
@@ -55,7 +60,9 @@ function getMatiere(req, res) {
     ]).then(response => {
         res.send(response)
     }).catch(err => {
-        res.send(err);
+        res.status(500).json({
+            message: "Une erreur est survenue, veuillez réessayer"
+        })
     });
 }
 
@@ -66,7 +73,9 @@ function updateMatiere(req, res) {
         { new: true },
         (err, matiere) => {
             if (err) {
-                res.json({ message: "Une erreur est survenue, modification non effectuée" });
+                res.status(500).json({
+                    message: "Une erreur est survenue, modification non effectuée"
+                })
             } else {
                 res.json({ message: "Modification effectuée" });
             }
@@ -75,12 +84,22 @@ function updateMatiere(req, res) {
 }
 
 function deleteMatiere(req, res) {
-    Matiere.findByIdAndRemove(req.params.id, (err, matiere) => {
-        if (err) {
-            res.send(err);
+    assignment.exists({ matiere: req.params.id }).then(response => {
+        if (response) {
+            res.status(409).json({
+                message: "Vous ne pouvez pas supprimé une matière lié à un assignment"
+            })
+        } else {
+            Matiere.findByIdAndRemove(req.params.id, (err, matiere) => {
+                if (err) {
+                    res.status(500).json({
+                        message: "Une erreur est survenue, veuillez réessayer"
+                    })
+                }
+                res.json({ message: `${matiere.libelle} supprimé` });
+            });
         }
-        res.json({ message: `${matiere.libelle} supprimé` });
-    });
+    })
 }
 
 function addMatiere(req, res) {
@@ -90,7 +109,9 @@ function addMatiere(req, res) {
 
     matiere.save((err) => {
         if (err) {
-            res.send("Erreur lors de la creation de la matiere:  ", err);
+            res.status(500).json({
+                message: "Une erreur est survenue, veuillez réessayer"
+            })
         }
         res.json({ message: `${matiere.libelle} enregistré avec succes !` });
     });
